requirements.txt
bardapi==0.1.29
flask
python-dotenv
